requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.2
python-dateutil==2.8.2
requests==2.31.0
urllib3==2.0.7
gunicorn==21.2.0
werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.9.0
